Shaped and TadukooButton moved to Tadukoo View
- Also they're properly Javadoc'd now
- Part of Issue #46 (need JUnit testing yet for TadukooButton)
