--- TadukooLookAndFeel/src/com/github/tadukoo/util/components/Shaped.java
@@ -1,10 +1,0 @@
-package com.github.tadukoo.util.components;
-
-import com.github.tadukoo.util.view.shapes.ShapeInfo;
-
-public interface Shaped{
-	
-	ShapeInfo getShapeInfo();
-	
-	void setShapeInfo(ShapeInfo shapeInfo);
-}
